+++ src/Adapter/Filesystem/Decode.php
@@ -6,6 +6,7 @@ namespace Formal\ORM\Adapter\Filesystem;
 use Formal\ORM\{
     Definition\Aggregate as Definition,
     Raw\Aggregate,
+    Raw\Aggregate\Collection\Entity\Reference,
 };
 use Innmind\Filesystem\{
     File,
@@ -129,10 +130,13 @@ final class Decode
                             static fn($collection) => Aggregate\Collection::of(
                                 $collection->name()->toString(),
                                 Set::of(...Json::decode($collection->content()->toString()))->map(
-                                    static fn($entity) => Set::of(...$entity)->map(
-                                        static fn($property) => Aggregate\Property::of(
-                                            $property[0],
-                                            $property[1],
+                                    static fn($entity) => Aggregate\Collection\Entity::of(
+                                        Reference::of($entity['reference']),
+                                        Set::of(...$entity['properties'])->map(
+                                            static fn($property) => Aggregate\Property::of(
+                                                $property[0],
+                                                $property[1],
+                                            ),
                                         ),
                                     ),
                                 ),