+++ src/Repository/Denormalize.php
@@ -6,7 +6,6 @@ namespace Formal\ORM\Repository;
 use Formal\ORM\{
     Definition\Aggregate as Definition,
     Definition\Aggregate\Property,
-    Definition\Aggregate\Entity,
     Raw\Aggregate,
     Id,
 };
@@ -40,8 +39,10 @@ final class Denormalize
     /**
      * @param Definition<T> $definition
      */
-    private function __construct(Definition $definition)
-    {
+    private function __construct(
+        Definition $definition,
+        KnownCollectionEntity $knownCollectionEntity,
+    ) {
         $this->definition = $definition;
         $this->instanciate = new Instanciate;
         /** @var \Closure(Aggregate\Id): Id<T> */
@@ -76,6 +77,7 @@ final class Denormalize
                 ->map(fn($collection) => [$collection->name(), Denormalize\Collection::of(
                     $collection,
                     $this->instanciate,
+                    $knownCollectionEntity,
                 )])
                 ->toList(),
         );
@@ -88,14 +90,14 @@ final class Denormalize
      */
     public function __invoke(Id $id = null): callable
     {
-        $id = match ($id) {
+        $denormalize = match ($id) {
             null => $this->denormalizeId,
             default => static fn(Aggregate\Id $_) => $id,
         };
 
         return fn(Aggregate $data) => Denormalized::of(
-            $id($data->id()),
-            $this->properties($data),
+            $id = $denormalize($data->id()),
+            $this->properties($id, $data),
         );
     }
 
@@ -107,15 +109,17 @@ final class Denormalize
      *
      * @return self<A>
      */
-    public static function of(Definition $definition): self
-    {
-        return new self($definition);
+    public static function of(
+        Definition $definition,
+        KnownCollectionEntity $knownCollectionEntity,
+    ): self {
+        return new self($definition, $knownCollectionEntity);
     }
 
     /**
      * @return Map<non-empty-string, mixed>
      */
-    private function properties(Aggregate $data): Map
+    private function properties(Id $id, Aggregate $data): Map
     {
         return Map::of(
             ...$data
@@ -160,7 +164,7 @@ final class Denormalize
                     fn($collection) => $this
                         ->collections
                         ->get($collection->name())
-                        ->map(static fn($denormalize): Set => $denormalize($collection))
+                        ->map(static fn($denormalize): Set => $denormalize($id, $collection))
                         ->map(static fn($value) => [$collection->name(), $value])
                         ->toSequence()
                         ->toSet(),