+++ src/Repository/Denormalize/Collection.php
@@ -7,6 +7,8 @@ use Formal\ORM\{
     Definition\Aggregate\Collection as Definition,
     Definition\Aggregate\Property,
     Raw\Aggregate\Collection as Raw,
+    Repository\KnownCollectionEntity,
+    Id,
 };
 use Innmind\Reflection\Instanciate;
 use Innmind\Immutable\{
@@ -23,16 +25,21 @@ final class Collection
     /** @var Definition<T> */
     private Definition $definition;
     private Instanciate $instanciate;
+    private KnownCollectionEntity $knownCollectionEntity;
     /** @var Map<non-empty-string, Property<T, mixed>> */
     private Map $properties;
 
     /**
      * @param Definition<T> $definition
      */
-    private function __construct(Definition $definition, Instanciate $instanciate)
-    {
+    private function __construct(
+        Definition $definition,
+        Instanciate $instanciate,
+        KnownCollectionEntity $knownCollectionEntity,
+    ) {
         $this->definition = $definition;
         $this->instanciate = $instanciate;
+        $this->knownCollectionEntity = $knownCollectionEntity;
         $this->properties = Map::of(
             ...$definition
                 ->properties()
@@ -44,15 +51,24 @@ final class Collection
     /**
      * @return Set<T>
      */
-    public function __invoke(Raw $collection): Set
+    public function __invoke(Id $id, Raw $collection): Set
     {
         $class = $this->definition->class();
+        $name = $collection->name();
+        // We use a weak reference to the aggregate id otherwise the closure below
+        // will always keep a reference to the id and all the weak maps that depend
+        // on this id will keep the associated data.
+        // This means that when reading lots of data the memory will always increase.
+        $idReference = \WeakReference::create($id);
 
         return $collection
-            ->properties()
-            ->map(function($properties) use ($class) {
-                $properties = Map::of(
-                    ...$properties
+            ->newEntities()
+            ->map(function($entity) use ($class, $idReference, $name) {
+                $reference = $entity->reference();
+
+                $entity = Map::of(
+                    ...$entity
+                        ->properties()
                         ->flatMap(
                             fn($property) => $this
                                 ->properties
@@ -66,10 +82,17 @@ final class Collection
                 );
 
                 /** @var T */
-                return ($this->instanciate)($class, $properties)->match(
-                    static fn($collection) => $collection,
-                    static fn() => throw new \RuntimeException("Unable to denormalize collection of type '$class'"),
-                );
+                return ($this->instanciate)($class, $entity)
+                    ->map(fn($object) => $this->knownCollectionEntity->add(
+                        $idReference,
+                        $name,
+                        $object,
+                        $reference,
+                    ))
+                    ->match(
+                        static fn($object) => $object,
+                        static fn() => throw new \RuntimeException("Unable to denormalize collection of type '$class'"),
+                    );
             });
     }
 
@@ -81,8 +104,11 @@ final class Collection
      *
      * @return self<A>
      */
-    public static function of(Definition $definition, Instanciate $instanciate): self
-    {
-        return new self($definition, $instanciate);
+    public static function of(
+        Definition $definition,
+        Instanciate $instanciate,
+        KnownCollectionEntity $knownCollectionEntity,
+    ): self {
+        return new self($definition, $instanciate, $knownCollectionEntity);
     }
 }