+++ src/Repository/Normalize.php
@@ -32,8 +32,10 @@ final class Normalize
     /**
      * @param Definition<T> $definition
      */
-    private function __construct(Definition $definition)
-    {
+    private function __construct(
+        Definition $definition,
+        KnownCollectionEntity $knownCollectionEntity,
+    ) {
         $this->definition = $definition;
         $this->extract = new Extract;
         $this->normalizeEntity = Map::of(
@@ -60,6 +62,7 @@ final class Normalize
                 ->map(fn($collection) => [$collection, Normalize\Collection::of(
                     $collection,
                     $this->extract,
+                    $knownCollectionEntity,
                 )])
                 ->toList(),
         );
@@ -72,6 +75,7 @@ final class Normalize
     {
         $properties = $denormalized->properties();
 
+        /** @psalm-suppress MixedArgument Due to the collection normalization */
         return Aggregate::of(
             $this->definition->id()->normalize($denormalized->id()),
             $this
@@ -127,7 +131,7 @@ final class Normalize
                         ->flatMap(
                             static fn($normalize) => $properties
                                 ->get($collection->name())
-                                ->map($normalize),
+                                ->map(static fn($object) => $normalize($denormalized->id(), $object)),
                         )
                         ->toSequence()
                         ->toSet(),
@@ -143,8 +147,10 @@ final class Normalize
      *
      * @return self<A>
      */
-    public static function of(Definition $definition): self
-    {
-        return new self($definition);
+    public static function of(
+        Definition $definition,
+        KnownCollectionEntity $knownCollectionEntity,
+    ): self {
+        return new self($definition, $knownCollectionEntity);
     }
 }