+++ src/Repository/Normalize/Collection.php
@@ -7,11 +7,13 @@ use Formal\ORM\{
     Definition\Aggregate\Collection as Definition,
     Raw\Aggregate\Collection as Raw,
     Raw\Aggregate\Property,
+    Repository\KnownCollectionEntity,
+    Id,
 };
 use Innmind\Reflection\Extract;
 use Innmind\Immutable\{
     Set,
-    Maybe,
+    Map,
 };
 
 /**
@@ -23,16 +25,21 @@ final class Collection
     /** @var Definition<T> */
     private Definition $definition;
     private Extract $extract;
+    private KnownCollectionEntity $knowCollectionEntity;
     /** @var Set<non-empty-string> */
     private Set $properties;
 
     /**
      * @param Definition<T> $definition
      */
-    private function __construct(Definition $definition, Extract $extract)
-    {
+    private function __construct(
+        Definition $definition,
+        Extract $extract,
+        KnownCollectionEntity $knownCollectionEntity,
+    ) {
         $this->definition = $definition;
         $this->extract = $extract;
+        $this->knowCollectionEntity = $knownCollectionEntity;
         $this->properties = $definition
             ->properties()
             ->map(static fn($property) => $property->name());
@@ -41,33 +48,49 @@ final class Collection
     /**
      * @param Set<T> $collection
      */
-    public function __invoke(Set $collection): Raw
+    public function __invoke(Id $id, Set $collection): Raw
     {
         $class = $this->definition->class();
-        $properties = $collection->map(
-            fn($object) => ($this->extract)($object, $this->properties)->match(
-                static fn($properties) => $properties,
+        $entities = Map::of(
+            ...$collection
+                ->map(fn($object) => [
+                    $this->knowCollectionEntity->reference(
+                        $id,
+                        $this->definition->name(),
+                        $object,
+                    ),
+                    $object,
+                ])
+                ->toList(),
+        );
+        $entities = $entities->map(
+            fn($_, $object) => ($this->extract)($object, $this->properties)->match(
+                static fn($entity) => $entity,
                 static fn() => throw new \LogicException("Failed to extract properties from '$class'"),
             ),
         );
 
         return Raw::of(
             $this->definition->name(),
-            $properties->map(
-                fn($properties) => $this
-                    ->definition
-                    ->properties()
-                    ->flatMap(
-                        static fn($property) => $properties
-                            ->get($property->name())
-                            ->map(static fn($value) => Property::of(
-                                $property->name(),
-                                $property->type()->normalize($value),
-                            ))
-                            ->toSequence()
-                            ->toSet(),
-                    ),
-            ),
+            $entities
+                ->map(
+                    fn($_, $entity) => $this
+                        ->definition
+                        ->properties()
+                        ->flatMap(
+                            static fn($property) => $entity
+                                ->get($property->name())
+                                ->map(static fn($value) => Property::of(
+                                    $property->name(),
+                                    $property->type()->normalize($value),
+                                ))
+                                ->toSequence()
+                                ->toSet(),
+                        ),
+                )
+                ->map(Raw\Entity::of(...))
+                ->values()
+                ->toSet(),
         );
     }
 
@@ -79,8 +102,11 @@ final class Collection
      *
      * @return self<A>
      */
-    public static function of(Definition $definition, Extract $extract): self
-    {
-        return new self($definition, $extract);
+    public static function of(
+        Definition $definition,
+        Extract $extract,
+        KnownCollectionEntity $knownCollectionEntity,
+    ): self {
+        return new self($definition, $extract, $knownCollectionEntity);
     }
 }